FEAT: new issue 페이지 코멘트 부분 제외 구현 완료(#45)
- 코멘트 text area부분 제외하고, 나머지 컴포넌트를 이용해서 new issue 페이지를 제작하였습니다.
